requirements.txt
mysql-connector-python==8.0.29
python-dotenv==0.20.0

